+++ src/agentscope/service/__init__.py
@@ -61,7 +61,6 @@ from .web.wikipedia import (
 )
 
 from .browser.web_browser import WebBrowser, WebElementInfo
-from .auto_service_toolkit import AutoServiceToolkit
 
 
 def get_help() -> None:
@@ -77,7 +76,6 @@ __all__ = [
     "ServiceResponse",
     "ServiceExecStatus",
     "ServiceToolkit",
-    "AutoServiceToolkit",
     "get_help",
     "execute_python_code",
     "execute_shell_command",
